Fix: ImagePicker 구조체 모델 분리 및 이름 변경
- MetadataLoader : PHAsset에서 메타데이터를 로드
- ImageLoader : 이미지 파일을 로드하고 해당 이미지를 UI에 표시
- DateFormatter : 날짜 형식화(원하는 형식으로 변경, 문자열을 Date 객체로 변환)
- LocationInfoProcessor : GPS 좌표를 사용하여 위치 정보를 처리
- LoadMetaDataView : 이미지 에셋 관리 뷰
