+++ FiveGuyes/FiveGuyes/FiveGuyesApp.swift
@@ -11,7 +11,8 @@ import SwiftUI
 struct FiveGuyesApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+//            ContentView()
+            TestingMetaView()
         }
     }
 }
